clean testcode
